testar transferencia de material

+++ conferencia-api/src/main/java/br/com/rpmont/conferencia/model/NotificacaoFenoRacao.java
@@ -84,5 +84,5 @@ public class NotificacaoFenoRacao {
     private LocalDateTime dataLeitura;
 
     @Column(name = "`usuarioLeituraID`")
-    private Integer usuarioLeituraId;
+    private Long usuarioLeituraId;
 }
